Rename GraphQl to Graph
... to conform to how we call it `Pouch` and not `PouchDb`
