+++ packages/db/src/meta/definitions.ts
@@ -2,7 +2,7 @@ import { logger } from "@truffle/db/logger";
 const debug = logger("db:meta:definitions");
 
 import { Collections, CollectionName } from "./collections";
-import * as GraphQl from "./graphql";
+import * as Graph from "./graph";
 import * as Pouch from "./pouch";
 import * as Process from "./process";
 
@@ -13,7 +13,7 @@ import * as Process from "./process";
  */
 export type Definitions<C extends Collections> = {
   [N in CollectionName<C>]: Definition<C, N>;
-  // [N in CollectionName<C>]: GraphQl.Definition<C, N> &
+  // [N in CollectionName<C>]: Graph.Definition<C, N> &
   //   Process.Definition<C, N> &
   //   Pouch.Definition<C, N>;
 };
@@ -22,7 +22,7 @@ export type Definitions<C extends Collections> = {
  * Main Definition type for whole meta system
  *
  * Each definition for a given [[CollectionName]] must adhere to the
- * [[GraphQl.Definition]], [[Pouch.Definition]], and [[Process.Definition]]
+ * [[Graph.Definition]], [[Pouch.Definition]], and [[Process.Definition]]
  * definition type for that name.
  *
  * @category Definitions
@@ -30,7 +30,7 @@ export type Definitions<C extends Collections> = {
 export type Definition<
   C extends Collections,
   N extends CollectionName<C>
-> = GraphQl.Definition<C, N> &
+> = Graph.Definition<C, N> &
   Process.Definition<C, N> &
   Pouch.Definition<C, N>;
 // > = Definitions<C>[N];
